+++ packages/Button/stories/OutlinedButton.stories.tsx
@@ -1,27 +1,15 @@
 import { Meta, Story } from '@storybook/react/types-6-0';
 import React, { useState } from 'react';
-import { BlazeButton } from '../src/BlazeButton';
-import { adminTheme, frontendTheme, ThemeContext, applyTheme } from '@blaze-react/themes';
+import Button from '../src';
+import { adminTheme, frontendTheme, applyTheme } from '@blaze-react/themes';
+import { args, argTypes } from './shared';
+import CustomIcon from '@blaze-react/icon';
 
 export default {
-  title: '@blaze-react/Button/Outlined',
-  component: BlazeButton,
-  args: {
-    selectedTheme: 'admin',
-  },
-  argTypes: {
-    selectedTheme: {
-      options: ['admin', 'frontend'],
-      control: { type: 'radio' },
-      defaultValue: { summary: 'admin' },
-    },
-    theme: {
-      control: false,
-      table: {
-        disable: true,
-      },
-    },
-  },
+  title: '@blaze-react/Button/All Stories/Outlined',
+  component: Button,
+  args: args,
+  argTypes: argTypes,
 } as Meta;
 
 const Template: Story = (args) => {
@@ -31,9 +19,31 @@ const Template: Story = (args) => {
   applyTheme(selectedTheme);
 
   return (
-    <ThemeContext.Provider value={{ theme, setTheme }}>
-      <BlazeButton {...args}>Blaze</BlazeButton>
-    </ThemeContext.Provider>
+    <Button {...args}>
+      <div className="flex items-center justify-center">
+        {args.icon === 'left' ? (
+          <>
+            <CustomIcon utilities="w-6 h-6 stroke-current fill-current" content={args.svgIcon} />
+            {args.icon !== 'icon-only' && <span className="mx-1">{args.label ?? 'Blaze'}</span>}
+            {args.icon === 'icon-only' && args.icon === 'with-text' && (
+              <span className="mx-1">{args.label ?? 'Blaze'}</span>
+            )}
+          </>
+        ) : (
+          <>
+            {args.icon !== 'icon-only' && <span className="mx-1">{args.label ?? 'Blaze'}</span>}
+
+            {args.icon !== 'no-icon' && (
+              <CustomIcon utilities="w-6 h-6 stroke-current fill-current" content={args.svgIcon} />
+            )}
+
+            {args.icon === 'icon-only' && args.icon === 'with-text' && (
+              <span className="mx-1">{args.label ?? 'Blaze'}</span>
+            )}
+          </>
+        )}
+      </div>
+    </Button>
   );
 };
 
@@ -43,6 +53,8 @@ export const Small = Template.bind({});
 export const Medium = Template.bind({});
 export const Large = Template.bind({});
 export const Stretched = Template.bind({});
+export const IconLeft = Template.bind({});
+export const IconRight = Template.bind({});
 
 Outlined.args = {
   theme: adminTheme,
@@ -55,7 +67,7 @@ Outlined.args = {
     text-button-outlined
     px-6
     py-1
-    w-24
+    min-w-min
     dark:bg-gray-800
     dark:text-button-primary
     border-2
@@ -76,7 +88,7 @@ Disabled.args = {
     rounded-button
     px-6
     py-1
-    w-24 
+    min-w-min 
     dark:bg-gray-800
     dark:text-button-primary
     border-2
@@ -100,7 +112,7 @@ Small.args = {
     text-button-primary 
     px-5 
     py-1 
-    w-20
+    min-w-min
     border-2
     border-button-outlined
     hover:bg-button-outlined-hover
@@ -126,7 +138,7 @@ Medium.args = {
     text-button-primary 
     px-6 
     py-1 
-    w-24 
+    min-w-min 
     border-2
     border-button-outlined
     hover:bg-button-outlined-hover
@@ -152,7 +164,7 @@ Large.args = {
     text-button-primary 
     px-6 
     py-1 
-    w-32
+    min-w-min
     border-2
     border-button-outlined
     hover:bg-button-outlined-hover
@@ -161,12 +173,13 @@ Large.args = {
     dark:text-button-primary
     focus:outline-none 
     active:bg-button-pressed
-    text-md
+    text-lg
   `,
 };
 
 Stretched.args = {
   theme: adminTheme,
+  icon: 'no-icon',
   utilities: `
   font-manrope
   font-medium
@@ -188,5 +201,53 @@ Stretched.args = {
   focus:outline-none
   active:bg-button-pressed
   text-md
+  `,
+};
+
+IconLeft.args = {
+  icon: 'left',
+  theme: adminTheme,
+  utilities: `
+  font-manrope
+  font-medium
+  text-bold
+  rounded-button
+  bg-button-outlined
+  text-button-outlined
+  px-6
+  py-1
+  min-w-min
+  dark:bg-gray-800
+  dark:text-button-primary
+  border-2
+  border-button-outlined
+  hover:bg-button-outlined-hover
+  hover:text-button-primary
+  focus:outline-none
+  active:bg-button-pressed
+  `,
+};
+
+IconRight.args = {
+  icon: 'right',
+  theme: adminTheme,
+  utilities: `
+  font-manrope
+  font-medium
+  text-bold
+  rounded-button
+  bg-button-outlined
+  text-button-outlined
+  px-6
+  py-1
+  min-w-min
+  dark:bg-gray-800
+  dark:text-button-primary
+  border-2
+  border-button-outlined
+  hover:bg-button-outlined-hover
+  hover:text-button-primary
+  focus:outline-none
+  active:bg-button-pressed
   `,
 };