+++ packages/Icon/src/types.ts
@@ -1,0 +1,5 @@
+export interface IconProps {
+  children?: unknown;
+  content?: string;
+  utilities?: string;
+}